api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites
attrs
certifi
charset-normalizer
colorama
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
idna
iniconfig
numpy
packaging
pandas
pluggy
py
PyJWT==2.1.0
pytest
pytest-django
pytest-pythonpath
python-dateutil
python-dotenv
pytz==2020.1
requests
six
sqlparse==0.3.1
toml
urllib3
